Hackery to get a first cut of the migration done
Link the review to all of the part 2 forms.
Update all the submissions with the project ID after the reviews are created.
Add some code to get the list of existing submissions so we don't have to recreate them if the script fails part way through.
Await confirmClearTable() in dump-all-fields.js.